KTOR-4157 Keep binary compatibility in CORS plugin
