--- ktor-server/ktor-server-plugins/ktor-server-cors-common/api/ktor-server-cors-common.api
@@ -1,44 +1,0 @@
-public final class io/ktor/server/plugins/cors/common/CORSCommonKt {
-	public static final fun buildCORSPlugin (Lio/ktor/server/application/PluginBuilder;)V
-}
-
-public final class io/ktor/server/plugins/cors/common/CORSConfig {
-	public static final field CORS_DEFAULT_MAX_AGE J
-	public static final field Companion Lio/ktor/server/plugins/cors/common/CORSConfig$Companion;
-	public fun <init> ()V
-	public final fun allowHeader (Ljava/lang/String;)V
-	public final fun allowHeaders (Lkotlin/jvm/functions/Function1;)V
-	public final fun allowHeadersPrefixed (Ljava/lang/String;)V
-	public final fun allowHost (Ljava/lang/String;Ljava/util/List;Ljava/util/List;)V
-	public static synthetic fun allowHost$default (Lio/ktor/server/plugins/cors/common/CORSConfig;Ljava/lang/String;Ljava/util/List;Ljava/util/List;ILjava/lang/Object;)V
-	public final fun allowMethod (Lio/ktor/http/HttpMethod;)V
-	public final fun allowXHttpMethodOverride ()V
-	public final fun anyHost ()V
-	public final fun exposeHeader (Ljava/lang/String;)V
-	public final fun getAllowCredentials ()Z
-	public final fun getAllowNonSimpleContentTypes ()Z
-	public final fun getAllowSameOrigin ()Z
-	public final fun getExposedHeaders ()Ljava/util/Set;
-	public final fun getHeaderPredicates ()Ljava/util/List;
-	public final fun getHeaders ()Ljava/util/Set;
-	public final fun getHosts ()Ljava/util/Set;
-	public final fun getMaxAgeInSeconds ()J
-	public final fun getMethods ()Ljava/util/Set;
-	public final fun setAllowCredentials (Z)V
-	public final fun setAllowNonSimpleContentTypes (Z)V
-	public final fun setAllowSameOrigin (Z)V
-	public final fun setMaxAgeInSeconds (J)V
-}
-
-public final class io/ktor/server/plugins/cors/common/CORSConfig$Companion {
-	public final fun getCorsDefaultMethods ()Ljava/util/Set;
-	public final fun getCorsSimpleContentTypes ()Ljava/util/Set;
-	public final fun getCorsSimpleRequestHeaders ()Ljava/util/Set;
-	public final fun getCorsSimpleResponseHeaders ()Ljava/util/Set;
-}
-
-public final class io/ktor/server/plugins/cors/common/KotlinTimeJvmKt {
-	public static final fun getMaxAgeDuration (Lio/ktor/server/plugins/cors/common/CORSConfig;)J
-	public static final fun setMaxAgeDuration-HG0u8IE (Lio/ktor/server/plugins/cors/common/CORSConfig;J)V
-}
-
