--- ktor-server/ktor-server-plugins/ktor-server-cors-routing/api/ktor-server-cors-routing.api
@@ -1,4 +1,0 @@
-public final class io/ktor/server/plugins/cors/routing/CORSRouteScopedKt {
-	public static final fun getCORSRouteScoped ()Lio/ktor/server/application/RouteScopedPlugin;
-}
-
