+++ ktor-server/ktor-server-plugins/ktor-server-cors/api/ktor-server-cors.api
@@ -1,4 +1,48 @@
+public final class io/ktor/server/plugins/cors/CORSConfig {
+	public static final field CORS_DEFAULT_MAX_AGE J
+	public static final field Companion Lio/ktor/server/plugins/cors/CORSConfig$Companion;
+	public fun <init> ()V
+	public final fun allowHeader (Ljava/lang/String;)V
+	public final fun allowHeaders (Lkotlin/jvm/functions/Function1;)V
+	public final fun allowHeadersPrefixed (Ljava/lang/String;)V
+	public final fun allowHost (Ljava/lang/String;Ljava/util/List;Ljava/util/List;)V
+	public static synthetic fun allowHost$default (Lio/ktor/server/plugins/cors/CORSConfig;Ljava/lang/String;Ljava/util/List;Ljava/util/List;ILjava/lang/Object;)V
+	public final fun allowMethod (Lio/ktor/http/HttpMethod;)V
+	public final fun allowXHttpMethodOverride ()V
+	public final fun anyHost ()V
+	public final fun exposeHeader (Ljava/lang/String;)V
+	public final fun getAllowCredentials ()Z
+	public final fun getAllowNonSimpleContentTypes ()Z
+	public final fun getAllowSameOrigin ()Z
+	public final fun getExposedHeaders ()Ljava/util/Set;
+	public final fun getHeaderPredicates ()Ljava/util/List;
+	public final fun getHeaders ()Ljava/util/Set;
+	public final fun getHosts ()Ljava/util/Set;
+	public final fun getMaxAgeInSeconds ()J
+	public final fun getMethods ()Ljava/util/Set;
+	public final fun setAllowCredentials (Z)V
+	public final fun setAllowNonSimpleContentTypes (Z)V
+	public final fun setAllowSameOrigin (Z)V
+	public final fun setMaxAgeInSeconds (J)V
+}
+
+public final class io/ktor/server/plugins/cors/CORSConfig$Companion {
+	public final fun getCorsDefaultMethods ()Ljava/util/Set;
+	public final fun getCorsSimpleContentTypes ()Ljava/util/Set;
+	public final fun getCorsSimpleRequestHeaders ()Ljava/util/Set;
+	public final fun getCorsSimpleResponseHeaders ()Ljava/util/Set;
+}
+
 public final class io/ktor/server/plugins/cors/CORSKt {
 	public static final fun getCORS ()Lio/ktor/server/application/ApplicationPlugin;
+}
+
+public final class io/ktor/server/plugins/cors/KotlinTimeJvmKt {
+	public static final fun getMaxAgeDuration (Lio/ktor/server/plugins/cors/CORSConfig;)J
+	public static final fun setMaxAgeDuration-HG0u8IE (Lio/ktor/server/plugins/cors/CORSConfig;J)V
+}
+
+public final class io/ktor/server/plugins/cors/routing/CORSKt {
+	public static final fun getCORS ()Lio/ktor/server/application/RouteScopedPlugin;
 }
