+++ ktor-server/ktor-server-plugins/ktor-server-cors/build.gradle.kts
@@ -3,11 +3,3 @@
  */
 
 description = ""
-
-kotlin.sourceSets {
-    commonMain {
-        dependencies {
-            api(project(":ktor-server:ktor-server-plugins:ktor-server-cors-common"))
-        }
-    }
-}
