+++ ktor-server/ktor-server-plugins/ktor-server-cors/jvmAndNix/src/io/ktor/server/plugins/cors/KotlinTimeJvm.kt
@@ -2,7 +2,7 @@
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.plugins.cors.common
+package io.ktor.server.plugins.cors
 
 import kotlin.math.*
 import kotlin.time.*
