+++ settings.gradle.kts
@@ -98,9 +98,7 @@ include(":ktor-server:ktor-server-plugins:ktor-server-call-logging")
 include(":ktor-server:ktor-server-plugins:ktor-server-compression")
 include(":ktor-server:ktor-server-plugins:ktor-server-conditional-headers")
 include(":ktor-server:ktor-server-plugins:ktor-server-content-negotiation")
-include(":ktor-server:ktor-server-plugins:ktor-server-cors-common")
 include(":ktor-server:ktor-server-plugins:ktor-server-cors")
-include(":ktor-server:ktor-server-plugins:ktor-server-cors-routing")
 include(":ktor-server:ktor-server-plugins:ktor-server-data-conversion")
 include(":ktor-server:ktor-server-plugins:ktor-server-default-headers")
 include(":ktor-server:ktor-server-plugins:ktor-server-double-receive")
